pyproject.toml
[tool.pytest.ini_options]
addopts = "-s -v --tb=auto"
testpaths = "tests"

env = [
    "STAGE=testing",
    "SQLALCHEMY_WARN_20=1"
]


[tool.poetry]
name = "bering-assignment"
version = "1.0.0"
description = "bering-assignment"
authors = ["Migorithm"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
pytest = "*"
sqlalchemy = "*"
greenlet = "*"
httpx = "*"
uvloop = "*"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
pymysql = "^1.0.3"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
asyncmy = "^0.2.8"
asyncpg = "^0.27.0"
aiomysql = "^0.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "*"
mypy = "*"
coverage = "*"
Faker = "*"
pytest-asyncio = "*"
pytest = "*"
pytest-cov = "*"
pytest-env = "*"
pytest-dotenv = "*"
types-requests = "*"
debugpy = "*"
black = "^23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
